+++ tests/Util/ExitException.php
@@ -5,7 +5,7 @@
  * @package news-sitemap-for-publishers
  */
 
-namespace Alley\WP\News_Sitemap\Tests\Util;
+namespace Alley\WP\News_Sitemap_For_Publishers\Tests\Util;
 
 /**
  * ExitException class.